C API: Digital IO function use stdbool.h
This requirement is already in place and makes the API more consistent

+++ api/icsneolegacy/icsneolegacy.cpp
@@ -497,7 +497,7 @@ int icsneoEnableDOIPLine(void* hObject, bool enable) {
 	if(!icsneoValidateHObject(hObject))
 		return false;
 	neodevice_t* device = (neodevice_t*)hObject;
-	return icsneo_setDigitalIO(device, ICSNEO_IO_ETH_ACTIVATION, 1, uint8_t(enable));
+	return icsneo_setDigitalIO(device, ICSNEO_IO_ETH_ACTIVATION, 1, enable);
 }
 
 int icsneoStartSockServer(void* hObject, int iPort) {
